pyproject.toml
[project]
name = "playground-organizer"
version = "0.1.0"
description = "Playground Organizer - Automatically organize files by access frequency"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
playground-organizer = "backend.src.playground_organizer.playground_organizer:main"
web-ui = "backend.start_web_server:main"
web-ui-debug = "backend.start_web_server_debug:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
]
